support IPA/AD cross domain trust
This version adds support for doing the ipa-adtrust-install and doing
the ipa trust-add with each Windows machine.  The work is done in the
vm-post-cloud-init.sh script which is run as the final runcmd in the
cloud-init for the VM.  This script does all of the setup, and will
demonstrate the use of Windows credentials to do GSSAPI authentication
against IPA, and do a -Y GSSAPI ldapsearch against AD.

ipa-adtrust-install restarts dirsrv, so the script will sleep for a
while to give dirsrv a chance to restart, and for named to
reconnect and sync.

The virtual network uses 192.168.128 mask 255.255.224.0, which gives
the IP address range 192.168.128.1-192.168.159.255
Use test. as the top level domain, so as not to conflict with any
other networks.

There are two virtual machines defined - ipa.ipa1dom.test with
IP address 192.168.129.2, and ad.ad1dom.test with IP address
192.168.130.2.  These are defined in ipa1.conf and ad1.conf
respectively.  Additional machines can be defined by copy-pasting
ipa1.conf -> ipa2.conf, ad1.conf -> ad2.conf, etc. and running the script
like this:

PATH=$PATH:/path/setupvm.sh/isin:/path/make-ad-vm.sh/isin:. ./ipa-ad-setup.sh global.conf ipa1.conf .... ipaN.conf ad1.conf ... adN.conf

There seems to be a problem on F20 when using a windows backing disk image,
and running virt-win-reg on the new disk image that uses the backing image.
The new disk image registry is somehow corrupted, which causes the
VM_WAIT_FILE to fail, as virt-ls and virt-cat fail with strange hivex
registry corruption errors.  The solution is to not use the downloaded
image as a backing image, but instead to just make a copy of it.  This takes
longer but is much more robust.

This version assumes you have created a new backing disk image for the ipa
machine, based on the standard Fedora 20 cloud image, updated with the latest
packages, and installed with all of the packages needed for ipa cross domain
trust.  This is the f20-ipa.qcow2 image in the ipa1.conf file.

set VM_DEBUG=1 to get debug messages from the VM creation.
Use unar if available, otherwise, use unrar.